src/IdServer/requirements.txt
flask
pyjwt
pycrypto
pycryptodome
gunicorn
flasgger
flask-cors
mysql-connector-python
